requirements.txt
streamlit
jinja2
xhtml2pdf
arabic-reshaper
python-bidi
